🤏 Confirmation dialog before deleting a project

+++ src/Project/ProjectManager.cpp
@@ -87,9 +87,9 @@ void ProjectManager::imgui(std::function<void(Project const&)> const& launch_pro
         {
             if (ImGui::Selectable("Delete project"))
             {
+                if (boxer::Selection::OK == boxer::show("Are you sure? This cannot be undone", "Deleting project", boxer::Style::Warning, boxer::Buttons::OKCancel))
                 // TODO(Launcher) move to trash
                 // and move to our own "trash", so that we can CTRL+Z the deletion
-                // TODO(Launcher) boxer confirmation windows
                 Cool::File::remove_folder(project.info_folder_path());
                 Cool::File::remove_file(project.file_path());
                 project_to_remove = it;
